+++ source/brain/robotstatemachine.cpp
@@ -255,9 +255,9 @@ namespace brain{
     void CRobotStateMachine::serialCallbackVCDcommand(char const * message, char * response)
     {
         int speed, steer;
-        uint8_t time_deciseconds;
+        uint16_t time_deciseconds;
 
-        uint8_t parsed = sscanf(message, "%d;%d;%hhu", &speed, &steer, &time_deciseconds);
+        uint8_t parsed = sscanf(message, "%d;%d;%hu", &speed, &steer, &time_deciseconds);
 
         if(uint8_globalsV_value_of_kl != 30){
             sprintf(response,"kl 30 is required!!");
@@ -296,9 +296,9 @@ namespace brain{
     void CRobotStateMachine::serialCallbackVCDCalibcommand(char const * message, char * response)
     {
         int speed, steer;
-        uint8_t time_deciseconds;
+        uint16_t time_deciseconds;
 
-        uint8_t parsed = sscanf(message, "%d;%d;%hhu", &speed, &steer, &time_deciseconds);
+        uint8_t parsed = sscanf(message, "%d;%d;%hu", &speed, &steer, &time_deciseconds);
 
         if(uint8_globalsV_value_of_kl != 30){
             sprintf(response,"kl 30 is required!!");
